Move compilation code to Compiler.mod
